IntimateTales_Roleplay/tests/test_core_plugin.py
```python
import unittest
from src.integration.core_plugin import CorePlugin

class TestCorePlugin(unittest.TestCase):

    def setUp(self):
        self.core_plugin = CorePlugin()

    def test_integrate_with_core_plugin(self):
        result = self.core_plugin.integrate_with_core_plugin()
        self.assertEqual(result, PLUGIN_STATUS['INTEGRATED'])

    def test_core_plugin_data(self):
        data = self.core_plugin.core_plugin_data
        self.assertIsInstance(data, dict)

    def test_core_plugin_update(self):
        old_data = self.core_plugin.core_plugin_data
        self.core_plugin.core_plugin_update()
        new_data = self.core_plugin.core_plugin_data
        self.assertNotEqual(old_data, new_data)

if __name__ == '__main__':
    unittest.main()
```